refactor: 边框11 tsx 改造
