+++ packages/datav-vue3/components/BorderBox11/index.ts
@@ -1,5 +1,5 @@
 import type { App, Plugin } from 'vue'
-import BorderBox11 from './src/index.vue'
+import BorderBox11 from './src/BorderBox11'
 
 export const BorderBox11Plugin: Plugin = {
   install(app: App) {
