+++ packages/docs/docs/Border/BorderBox11/demo.vue
@@ -1,9 +1,12 @@
 <template>
   <div demo-bg>
-    <dv-border-box-11 title="dv-border-box-11" :title-width="400">
+    <dv-border-box11 title="dv-border-box-11" :title-width="400">
       <div dv-bg>
         dv-border-box-11
       </div>
-    </dv-border-box-11>
+    </dv-border-box11>
   </div>
 </template>
+<script setup lang="ts">
+import { BorderBox11 as DvBorderBox11 } from '@kjgl77/datav-vue3'
+</script>
